Added geting city from reversed geocoding api
